Fixing up the Windows CLI object thing
It can depend on the whole config. This is paving the way to
have the configurations inherit from the base one. I can't
get the chdir to work reliably (specifically, for attrib),
so I'm going to stick with creating the entire command string.

I feel like I could get away from that and remove the whole
cautious quoting.